requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
celery[redis]==5.4.0
redis==5.1.0
python-dotenv==1.0.1
google-generativeai==0.8.0
PyMuPDF==1.24.0
python-docx==1.1.2
Pillow==10.4.0
pytesseract==0.3.13
pydantic==2.9.0
gunicorn==22.0.0
